Player Lives & Health Restore Upon Player Death
I have succeed in making the players health reset to 100 upon there
death via water/lava, and have started the development  for a player
lives UI bar and tracker.

+++ Lost Island/Assets/Scripts/EnemyAI.cs
@@ -5,7 +5,7 @@ public class EnemyAI : MonoBehaviour
 {
 	float distance;
 	public Transform target;
-	public int damage = 5;
+	public float damage = 5f;
 	public float lookAtDistance = 25.0f;
 	public float attackRange = 15.0f;
 	public float moveSpeed = 5.0f;
